Move Connection to its own file
along with PeerConnection
Leave BusConnection and friends in bus.rb
