+++ lib/dbus.rb
@@ -26,6 +26,7 @@ require_relative "dbus/api_options"
 require_relative "dbus/auth"
 require_relative "dbus/bus"
 require_relative "dbus/bus_name"
+require_relative "dbus/connection"
 require_relative "dbus/data"
 require_relative "dbus/emits_changed_signal"
 require_relative "dbus/error"
